requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.2.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.6.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.3.0
Babel==2.14.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.9.0
cachetools==5.5.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.8
comm==0.2.0
contourpy==1.2.1
cycler==0.12.1
datasets==3.1.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
Django==4.2
django-connections==0.1
django-crispy-forms==1.14.0
evaluate==0.4.3
executing==2.0.1
faiss-cpu==1.12.0
fastjsonschema==2.19.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.53.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.9.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.67.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.26.2
idna==3.6
ipykernel==6.27.1
ipython==8.19.0
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
jiter==0.10.0
joblib==1.4.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.1
jupyter_server==2.12.1
jupyter_server_terminals==0.5.0
jupyterlab==4.2.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
keras==3.6.0
kiwisolver==1.4.5
langchain==0.3.27
langchain-core==0.3.74
langchain-openai==0.3.31
langchain-text-splitters==0.3.9
langsmith==0.4.16
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
multidict==6.1.0
multiprocess==0.70.16
namex==0.0.8
narwhals==1.25.2
nbclient==0.9.0
nbconvert==7.13.1
nbformat==5.9.2
nest-asyncio==1.5.8
notebook==7.2.0
notebook_shim==0.2.3
numpy==1.26.4
openai==1.101.0
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.0
orjson==3.11.2
outcome==1.3.0.post0
overrides==7.4.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.1.0
plotly==5.22.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
propcache==0.2.0
protobuf==4.25.5
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==18.1.0
pycparser==2.21
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.17.2
pyparsing==3.1.2
pypdf==6.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.1.1
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
redis==5.1.0
referencing==0.32.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.15.2
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
selenium==4.28.0
Send2Trash==1.8.2
six==1.16.0
smmap==5.0.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
spotipy==2.24.0
SQLAlchemy==2.0.43
sqlparse==0.4.4
stack-data==0.6.3
streamlit==1.42.0
tenacity==8.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
terminado==0.18.0
threadpoolctl==3.5.0
tiktoken==0.11.0
tinycss2==1.2.1
tokenizers==0.20.3
toml==0.10.2
tornado==6.4
tqdm==4.67.0
traitlets==5.14.0
transformers==4.46.2
trio==0.28.0
trio-websocket==0.11.1
types-python-dateutil==2.8.19.14
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2023.3
uri-template==1.3.0
urllib3==2.1.0
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.5.0
yarl==1.18.0
zstandard==0.24.0
